revert to REST
DATA API started returning html body for a comment instead of the text
body. This broke everything in this app. This is a quick fix that uses
the rEST api to accomplish was the DATA API did. This also makes the
refactoring done for the DATA API wasteful and inefficient, yet
unnoticeable given the small amount of processing needed in the first
place.